requirements.txt
amqp==2.6.1
asgiref==3.3.4
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==4.4.7
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.3
django-bootstrap-form==3.4
django-crispy-forms==1.12.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-elasticsearch-dsl==7.2.0
django-embed-video==1.4.0
django-filter==2.4.0
django-google-analytics-app==5.0.3
django-humanize==0.1.2
django-summernote==0.8.11.6
django-tawkto==0.3
django-widget-tweaks==1.4.8
elasticsearch==7.13.1
elasticsearch-dsl==7.3.0
humanize==3.5.0
idna==2.10
kombu==4.6.11
Pillow==8.2.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
structlog==21.1.0
urllib3==1.26.4
vine==1.3.0
whitenoise==5.2.0
